+++ src/main/java/nl/inl/blacklab/perdocument/DocResult.java
@@ -21,6 +21,7 @@ import org.apache.lucene.document.Document;
 
 import nl.inl.blacklab.search.Hit;
 import nl.inl.blacklab.search.Hits;
+import nl.inl.blacklab.search.HitsImpl;
 import nl.inl.blacklab.search.Searcher;
 
 /**
@@ -43,7 +44,7 @@ public class DocResult {
 		this.docId = docId;
 		this.document = document;
 		this.score = score;
-		hits = new Hits(searcher, concField);
+		hits = new HitsImpl(searcher, concField);
 	}
 
 	/**
@@ -62,7 +63,7 @@ public class DocResult {
 		this.docId = doc;
 		this.document = document;
 		this.score = 0.0f;
-		hits = new Hits(searcher, concField, docHits);
+		hits = new HitsImpl(searcher, concField, docHits);
 	}
 
 	/**