+++ src/main/java/nl/inl/blacklab/search/grouping/HitGroup.java
@@ -19,6 +19,7 @@ import java.util.List;
 
 import nl.inl.blacklab.search.Hit;
 import nl.inl.blacklab.search.Hits;
+import nl.inl.blacklab.search.HitsImpl;
 import nl.inl.blacklab.search.Searcher;
 
 /**
@@ -30,7 +31,7 @@ public class HitGroup extends Group {
 
 	HitGroup(Searcher searcher, HitPropValue groupIdentity, String defaultConcField) {
 		super(groupIdentity);
-		results = new Hits(searcher, defaultConcField);
+		results = new HitsImpl(searcher, defaultConcField);
 	}
 
 	/**
@@ -45,7 +46,7 @@ public class HitGroup extends Group {
 	 */
 	HitGroup(Searcher searcher, HitPropValue groupIdentity, String defaultConcField, List<Hit> hits) {
 		super(groupIdentity);
-		results = new Hits(searcher, defaultConcField, hits);
+		results = new HitsImpl(searcher, defaultConcField, hits);
 	}
 
 	public Hits getHits() {