+++ src/site/markdown/changelog.md
@@ -12,7 +12,8 @@
 ### Changed
 * Performed some code cleanups, moved some internal classes to different packages.
 * Made Searcher an abstract base class to SearcherImpl.
-* Added mock classes for Searcher, ForwardIndex, Terms; used them to add tests.
+* Made Hits an abstract base class to HitsImpl.
+* Added mock classes for Searcher, Hits, ForwardIndex, Terms; used them to add tests.
 
 ## Improvements up to v1.3.5